+++ graceful/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be released
 
+* build(deps): update `github.com/Scalingo/go-utils/errors` from v2 to v3
+
 ## v1.3.1
 
 * chore(go): corrective bump - Go version regression from 1.24.3 to 1.24